refactor: HomeRoutineManager 데이터 연동 안 되는 문제 확인 -> HomeViewModel로 합침(SCRUM-225)
